# api-docs_Grupo_20.json
{"openapi":"3.1.0","info":{"title":"API de Donde Estás","description":"Documentación de la API REST del sistema de trazabilidad","contact":{"name":"Grupo 20","email":"grupo20@example.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/usuarios/{id}":{"get":{"tags":["Usuarios"],"operationId":"getUsuarioById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioSmallDTO"}}}}}},"put":{"tags":["Usuarios"],"operationId":"updateUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioSmallDTO"}}}}}}},"/api/ubicaciones/{id}":{"get":{"tags":["Ubicaciones"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UbicacionDTO"}}}}}},"put":{"tags":["Ubicaciones"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UbicacionUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/publicaciones/{id}":{"get":{"tags":["Publicaciones"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublicacionDTO"}}}}}},"put":{"tags":["Publicaciones"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicacionUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublicacionDTO"}}}}}},"delete":{"tags":["Publicaciones"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublicacionDTO"}}}}}}},"/api/publicaciones":{"get":{"tags":["Publicaciones"],"operationId":"getFiltered","parameters":[{"name":"filter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/PublicacionFilterDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginateBaseDTOPublicacionDTO"}}}}}},"post":{"tags":["Publicaciones"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicacionCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublicacionDTO"}}}}}}},"/api/avistamientos":{"get":{"tags":["Avistamientos"],"operationId":"getFiltered_1","parameters":[{"name":"filter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/AvistamientoFilterDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AvistamientoDTO"}}}}}}},"post":{"tags":["Avistamientos"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvistamientoCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AvistamientoDTO"}}}}}}},"/api/auth/register":{"post":{"tags":["Auth"],"operationId":"createUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioCreateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthSessionDTO"}}}}}}},"/api/auth/login":{"post":{"tags":["Auth"],"operationId":"login","parameters":[{"name":"usuario","in":"header","required":true,"schema":{"type":"string"}},{"name":"password","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthSessionDTO"}}}}}}},"/api/usuarios":{"get":{"tags":["Usuarios"],"operationId":"getAllUsuarios","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UsuarioSmallDTO"}}}}}}}},"/api/ubicaciones/buscar":{"get":{"tags":["Ubicaciones"],"operationId":"getFiltered_2","parameters":[{"name":"idExterno","in":"query","required":false,"schema":{"type":"string"}},{"name":"provincia","in":"query","required":false,"schema":{"type":"string"}},{"name":"ciudad","in":"query","required":false,"schema":{"type":"string"}},{"name":"barrio","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UbicacionDTO"}}}}}}}},"/api/avistamientos/{id}":{"get":{"tags":["Avistamientos"],"operationId":"get_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AvistamientoDTO"}}}}}}}},"components":{"schemas":{"UsuarioUpdateDTO":{"type":"object","properties":{"nombre":{"type":"string","maxLength":100,"minLength":0},"apellido":{"type":"string","maxLength":100,"minLength":0},"email":{"type":"string","format":"email"},"nuevaUbicacionId":{"type":"integer","format":"int64"}}},"UsuarioSmallDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"apellido":{"type":"string"},"email":{"type":"string"},"rol":{"type":"string","enum":["ADMINISTRADOR","USUARIO_COMUN"]}}},"UbicacionUpdateDTO":{"type":"object","properties":{"idExterno":{"type":"string"},"provincia":{"type":"string"},"ciudad":{"type":"string"},"barrio":{"type":"string"},"latitud":{"type":"number","format":"double"},"longitud":{"type":"number","format":"double"}}},"PublicacionUpdateDTO":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"color":{"type":"string"},"especie":{"type":"string"},"raza":{"type":"string"},"tamanio":{"type":"string"},"agradecimientos":{"type":"array","items":{"type":"integer","format":"int64"}},"ubicacion":{"$ref":"#/components/schemas/UbicacionUpdateDTO"},"estado":{"type":"string","enum":["PERDIDO_PROPIO","PERDIDO_AJENO","RECUPERADO","ADOPTADO"]}}},"EstadoPublicacionDTO":{"type":"object","properties":{"estado":{"type":"string","enum":["PERDIDO_PROPIO","PERDIDO_AJENO","RECUPERADO","ADOPTADO"]},"fecha":{"type":"string","format":"date-time"}}},"PublicacionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"descripcion":{"type":"string"},"fecha":{"type":"string","format":"date-time"},"color":{"type":"string"},"especie":{"type":"string"},"raza":{"type":"string"},"tamanio":{"type":"string"},"ubicacion":{"$ref":"#/components/schemas/UbicacionDTO"},"usuarioId":{"type":"integer","format":"int64"},"estado":{"$ref":"#/components/schemas/EstadoPublicacionDTO"},"fotos":{"type":"array","items":{"type":"string"}}}},"UbicacionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"idExterno":{"type":"string"},"provincia":{"type":"string"},"ciudad":{"type":"string"},"barrio":{"type":"string"},"latitud":{"type":"number","format":"double"},"longitud":{"type":"number","format":"double"}}},"PublicacionCreateDTO":{"type":"object","properties":{"nombre":{"type":"string","minLength":1},"descripcion":{"type":"string"},"color":{"type":"string","minLength":1},"especie":{"type":"string","minLength":1},"raza":{"type":"string","minLength":1},"tamanio":{"type":"string","minLength":1},"ubicacion":{"$ref":"#/components/schemas/UbicacionCreateDTO"},"estado":{"type":"string","enum":["PERDIDO_PROPIO","PERDIDO_AJENO","RECUPERADO","ADOPTADO"]}},"required":["color","especie","estado","nombre","raza","tamanio","ubicacion"]},"UbicacionCreateDTO":{"type":"object","properties":{"idExterno":{"type":"string"},"provincia":{"type":"string","minLength":1},"ciudad":{"type":"string","minLength":1},"barrio":{"type":"string","minLength":1},"latitud":{"type":"number","format":"double"},"longitud":{"type":"number","format":"double"}},"required":["barrio","ciudad","latitud","longitud","provincia"]},"AvistamientoCreateDTO":{"type":"object","properties":{"descripcion":{"type":"string","minLength":1},"ubicacion":{"$ref":"#/components/schemas/UbicacionCreateDTO"},"publicacionId":{"type":"integer","format":"int64"}},"required":["descripcion","publicacionId","ubicacion"]},"AvistamientoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"descripcion":{"type":"string"},"agradecimiento":{"type":"boolean"},"fecha":{"type":"string","format":"date-time"},"ubicacion":{"$ref":"#/components/schemas/UbicacionDTO"},"publicacionId":{"type":"integer","format":"int64"},"usuario":{"$ref":"#/components/schemas/UsuarioSmallDTO"}}},"UsuarioCreateDTO":{"type":"object","properties":{"nombre":{"type":"string","minLength":1},"apellido":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6},"rol":{"type":"string","enum":["ADMINISTRADOR","USUARIO_COMUN"]},"ubicacion":{"$ref":"#/components/schemas/UbicacionCreateDTO"}},"required":["apellido","email","nombre","password","rol","ubicacion"]},"AuthSessionDTO":{"type":"object","properties":{"token":{"type":"string"},"usuario":{"$ref":"#/components/schemas/UsuarioSmallDTO"}}},"PublicacionFilterDTO":{"type":"object","properties":{"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"sortBy":{"type":"string"},"sortDir":{"type":"string"},"nombre":{"type":"string"},"especie":{"type":"string"},"raza":{"type":"string"},"tamanio":{"type":"string"},"color":{"type":"string"},"fechaDesde":{"type":"string","format":"date-time"},"fechaHasta":{"type":"string","format":"date-time"},"usuarioId":{"type":"integer","format":"int64"}}},"PaginateBaseDTOPublicacionDTO":{"type":"object","properties":{"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"elements":{"type":"array","items":{"$ref":"#/components/schemas/PublicacionDTO"}}}},"AvistamientoFilterDTO":{"type":"object","properties":{"publicacionId":{"type":"integer","format":"int64"},"usuarioId":{"type":"integer","format":"int64"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}